pyproject.toml
[project]
name = "aiapps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.40.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
]
